+++ examples/rally-api/src/integration-test/groovy/yakworks/rest/ApiResultsTestControllerTests.groovy
@@ -23,8 +23,21 @@ class ApiResultsTestControllerTests extends RestIntTest {
         then:
         body.ok == true
         body.status == 207
-        response.status == 207
+        //its not 207 by default anymore
+        response.status == 200
+    }
 
+    void "test get with problems"() {
+        when:
+        controller.getWithProblems()
+        Map body = response.bodyToMap()
+
+        then:
+        !body.ok
+        body.status == 207
+        body.title == "Unhandled Problem"
+        //its not 207 by default anymore
+        response.status == 200
     }
 
 }