+++ gorm-rest/src/main/groovy/yakworks/rest/gorm/controller/BulkControllerSupport.groovy
@@ -4,7 +4,7 @@
 */
 package yakworks.rest.gorm.controller
 
-import javax.inject.Inject
+
 import javax.servlet.http.HttpServletRequest
 
 import groovy.transform.CompileStatic
@@ -37,16 +37,16 @@ import yakworks.spring.AppCtx
 @SuppressWarnings(['CatchRuntimeException'])
 class BulkControllerSupport<D> {
 
-    @Autowired(required = false)
+    @Autowired
     SyncJobService syncJobService
 
-    @Autowired(required = false)
+    @Autowired
     CsvToMapTransformer csvToMapTransformer
 
-    @Autowired(required = false)
+    @Autowired
     IncludesConfig includesConfig
 
-    @Inject ProblemHandler problemHandler
+    @Autowired ProblemHandler problemHandler
 
     Class<D> entityClass // the domain class this is for
 
@@ -134,7 +134,7 @@ class BulkControllerSupport<D> {
      * Special handler for bulk operations, so that we can log/highight every bulk error we send.
      * Its here, because we cant have more thn one exception handler for "Exception" in controller
      */
-    Problem handleBulkOperationException(HttpServletRequest req, Exception e) {
+    Problem handleBulkOperationException(HttpServletRequest req, Throwable e) {
         Problem apiError = problemHandler.handleException(getEntityClass(), e)
         if (apiError.status.code == 500) {
             String requestInfo = "requestURI=[${req.requestURI}], method=[${req.method}], queryString=[${req.queryString}]"