+++ rally-domain/src/test/groovy/yakworks/rally/activity/ActivityContactSpec.groovy
@@ -12,10 +12,11 @@ import yakworks.rally.orgs.model.Contact
 import yakworks.rally.orgs.model.ContactSource
 import yakworks.rally.orgs.model.Org
 import yakworks.rally.tag.model.Tag
+import yakworks.testing.gorm.unit.GormHibernateTest
 import yakworks.testing.gorm.unit.SecurityTest
 import yakworks.testing.gorm.unit.DataRepoTest
 
-class ActivityContactSpec extends Specification implements DataRepoTest, SecurityTest {
+class ActivityContactSpec extends Specification implements GormHibernateTest, SecurityTest {
     static List entityClasses = [Activity, ActivityContact, ActivityNote, Contact, ContactSource, Org, Tag]
 
     @Shared
@@ -33,16 +34,16 @@ class ActivityContactSpec extends Specification implements DataRepoTest, Securit
 
     List createSomeContacts(Long orgId){
         List items = []
-        (0..9).each { eid ->
-            items <<  Contact.create(firstName: "Name$eid", org:[id: orgId])
+        (0..9).eachWithIndex { eid, idx ->
+            items <<  Contact.repo.create([id: idx+1, firstName: "Name$eid", org:[id: orgId]],[bindId:true])
         }
         flushAndClear()
         return items
     }
 
-    Activity setupData(List contacts = null){
-        def org = build(Org)
-        if(!contacts) contacts = createSomeContacts(org.id)
+    Activity setupData(List contacts){
+        // def org = build(Org)
+        // if(!contacts) contacts = createSomeContacts(org.id)
         List<Map> cmap = [
             [id:contacts[0].id],
             [id:contacts[1].id],
@@ -52,7 +53,7 @@ class ActivityContactSpec extends Specification implements DataRepoTest, Securit
         flush()
 
         def data = [
-            org    : [id: 1],
+            org    : contacts[0].org,
             note   : [body: 'test note'],
             contacts: cmap
         ]
@@ -67,9 +68,9 @@ class ActivityContactSpec extends Specification implements DataRepoTest, Securit
         when:
         def org = build(Org)
 
-        def c = createSomeContacts(org.id)
-        def con1 = c[0]
-        def con2 = c[1]
+        List c = createSomeContacts(org.id)
+        Contact con1 = c[0]
+        Contact con2 = c[1]
         def act = setupData(c)
         flush() //needed for the checks as it queries
 
@@ -88,13 +89,14 @@ class ActivityContactSpec extends Specification implements DataRepoTest, Securit
 
         when: 'remove 1 by key'
         repo.remove(act, con2)
-
+        flush()
         then:
         !repo.exists(act, con2)
         repo.count(act) == 2
 
         when: 'remove all by org'
         repo.remove(act)
+        flush()
 
         then:
         repo.count(act) == 0