+++ rally-domain/src/test/groovy/yakworks/rally/orgs/OrgCopierSpec.groovy
@@ -1,18 +1,17 @@
 package yakworks.rally.orgs
 
 import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.cache.support.SimpleCacheManager
 
-import yakworks.rally.config.OrgProps
-import yakworks.rally.mail.model.MailMessage
-import yakworks.rally.testing.OrgDimensionTesting
-import yakworks.testing.gorm.RepoTestData
-import yakworks.testing.gorm.unit.SecurityTest
-import yakworks.testing.gorm.unit.DataRepoTest
 import spock.lang.Specification
 import yakworks.rally.activity.ActivityCopier
 import yakworks.rally.activity.model.Activity
 import yakworks.rally.activity.model.ActivityLink
+import yakworks.rally.attachment.AttachmentSupport
+import yakworks.rally.attachment.model.Attachment
 import yakworks.rally.attachment.model.AttachmentLink
+import yakworks.rally.config.OrgProps
+import yakworks.rally.mail.model.MailMessage
 import yakworks.rally.orgs.model.Contact
 import yakworks.rally.orgs.model.ContactEmail
 import yakworks.rally.orgs.model.ContactFlex
@@ -27,21 +26,30 @@ import yakworks.rally.orgs.model.OrgMember
 import yakworks.rally.orgs.model.OrgSource
 import yakworks.rally.orgs.model.OrgTag
 import yakworks.rally.orgs.model.OrgTypeSetup
+import yakworks.rally.testing.OrgDimensionTesting
+import yakworks.testing.gorm.RepoTestData
+import yakworks.testing.gorm.unit.DataRepoTest
+import yakworks.testing.gorm.unit.GormHibernateTest
+import yakworks.testing.gorm.unit.SecurityTest
 
-class OrgCopierSpec extends Specification implements DataRepoTest, SecurityTest {
+class OrgCopierSpec extends Specification implements GormHibernateTest, SecurityTest {
     static List entityClasses = [
         Org, Contact, OrgFlex, OrgMember, OrgCalc, OrgSource, OrgTag, OrgInfo, OrgTypeSetup, Location, ContactPhone,
-        ContactEmail, ContactSource, ContactFlex, Activity, ActivityLink, AttachmentLink, MailMessage]
+        ContactEmail, ContactSource, ContactFlex, Activity, ActivityLink, AttachmentLink, MailMessage, Attachment
+    ]
+
+    static List springBeans = [ActivityCopier, OrgCopier, OrgProps, OrgDimensionService ]
 
     @Autowired OrgCopier orgCopier
-    @Autowired OrgDimensionService orgDimensionService
 
-    Closure doWithGormBeans(){ { ->
-        orgDimensionService(OrgDimensionService)
-        orgCopier(OrgCopier)
-        activityCopier(ActivityCopier)
-        orgProps(OrgProps)
-    }}
+    // Closure doWithGormBeans(){ { ->
+    //     orgDimensionService(OrgDimensionService)
+    //     orgCopier(OrgCopier)
+    //     activityCopier(ActivityCopier)
+    //     orgProps(OrgProps)
+    //     cacheManager(SimpleCacheManager)
+    //     attachmentSupport(AttachmentSupport)
+    // }}
 
     def "test copy"() {
         setup:
@@ -50,9 +58,9 @@ class OrgCopierSpec extends Specification implements DataRepoTest, SecurityTest 
         Org old = build(Org)
         // old.type = TestData.build(OrgType)
         // old.orgTypeId = old.type.id
-        old.calc = build(OrgCalc)
-        old.flex = build(OrgFlex)
-        old.info = build(OrgInfo)
+        old.calc = build(OrgCalc, id: old.id)
+        old.flex = build(OrgFlex, id: old.id)
+        old.info = build(OrgInfo, id: old.id)
         old.member = RepoTestData.build(OrgMember, [id: old.id, branch: build(Org), division: build(Org)])
 
         Location location = RepoTestData.build(Location)